added controllers for relevant entities and refactored test cases

+++ src/main/java/com/fdmgroup/JeffersonExpenseTracker/Service/CategoryService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.fdmgroup.JeffersonExpenseTracker.Dao.CategoryRepository;
 import com.fdmgroup.JeffersonExpenseTracker.Dao.UserRepository;
 import com.fdmgroup.JeffersonExpenseTracker.Exceptions.CategoryIdException;
+import com.fdmgroup.JeffersonExpenseTracker.Exceptions.ExpenseIdException;
 import com.fdmgroup.JeffersonExpenseTracker.Exceptions.UserIdException;
 import com.fdmgroup.JeffersonExpenseTracker.Model.Category;
 import com.fdmgroup.JeffersonExpenseTracker.Model.User;
@@ -48,7 +49,12 @@ public class CategoryService {
 	}
 
 	public void deleteById(int categoryId) {
-		this.categoryRepo.deleteById(categoryId);
-
+		
+		if (categoryRepo.existsById(categoryId)) {
+			this.categoryRepo.deleteById(categoryId);
+			return;
+		}
+		
+		throw new CategoryIdException("Must provide a valid categoryId for deleting");
 	}
 }
